backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.44
psycopg[binary]==3.2.13
pydantic-settings==2.6.0
requests==2.32.3
APScheduler==3.10.4
alembic==1.13.3
python-dateutil==2.9.0.post0

pytest==8.3.3
pytest-mock==3.14.0
httpx==0.27.2
